requirements.txt
discord.py
openai>=1.0.0
python-dotenv
PyYAML
